Validate GraphQL payload and tweak method calling convention.
Every payload sent to the GraphQL server must include structured
query text so make the first argument to all the relevant methods
be that string.

With the above in place star to strengthen enforcement of the
payload by the client - ensure that the requested payload type
matches the method by which the action was constructed.